fix(codeql): dead-code/style batch — 70 changed, 26 skipped
9 rule classes (issue #78, wave 2.4), behavior-preserving by construction:

- useless-cast-to-self: 33 changed / 13 skipped. Skips = <unknown type>
  sites (NWBrainEntity:66,75; NWCreature:3856; UniverseBuilder:313),
  object-typed Owner/Clone casts load-bearing for overload resolution
  (EffectsFactory:2883,4355,4375,5068; BaseRoom:41; Building:98,144;
  NWCreature:3339; BaseControl:836).
- missed-readonly-modifier: 23 changed / 0 skipped (ctor-only assignments;
  modifier order matched). Build is the guard.
- useless-assignment-to-local: 8 changed / 2 skipped (NWCreature:649 CodeQL
  false positive — field read by loop; Gate:52 feeds property getter).
- simplifiable-boolean: 2 changed (A ? B : false -> A && B).
- useless-tostring: 1 changed. unused-property-value: 1 changed (empty
  Leadership setter removed — no assignments exist). unused-collection:
  1 changed (never-read fNames dict removed).
- linq/missed-where: 0 changed / 10 skipped; linq/missed-select:
  0 changed / 1 skipped — restraint rule (loops not side-effect-free
  pure filter/collect; imperative preserved).

No tests added (behavior-preserving by construction; full suite is guard).
Verified: xbuild 0 errors; Tier B + Tier C green; CRLF preserved.
Closes #78.

+++ project/Effects/EffectsFactory.cs
@@ -931,7 +931,7 @@ namespace NWR.Effects
             } else {
                 NWCreature mon = (NWCreature)aExtItem.Contents[0];
                 aExtItem.CLSID = GlobalVars.iid_Mummy;
-                aExtItem.Weight = (((float)mon.Weight * 0.1f));
+                aExtItem.Weight = ((mon.Weight * 0.1f));
                 string cs = mon.GetDeclinableName(Number.nSingle, Case.cGenitive);
                 GlobalVars.nwrHost.ShowText(GlobalVars.nwrHost, BaseLocale.Format(RS.rs_MummyCreated, new object[]{ cs }));
             }
@@ -4969,7 +4969,7 @@ namespace NWR.Effects
                 return;
             }
 
-            ushort tid = (ushort)tile_id;
+            ushort tid = tile_id;
             if ((StaticData.dbPlaces[tid].Signs.Contains(PlaceFlags.psBackground))) {
                 aTile.Background = tid;
                 return;
